+++ web/src/beta/features/Editor/Visualizer/type.ts
@@ -1,4 +1,11 @@
 import { Block } from "@reearth/beta/features/Visualizer/Crust";
+import { SceneMode } from "@reearth/beta/types";
+import {
+  ImageBasedLighting,
+  LightProperty,
+  ShadowProperty,
+  ViewerProperty as CoreViewerProperty
+} from "@reearth/core";
 
 export interface Item {
   id: string;
@@ -7,3 +14,16 @@ export interface Item {
 }
 
 export type BlockType = Item & Pick<Block, "pluginId" | "extensionId">;
+
+export declare type ViewerProperty = {
+  scene?: {
+    backgroundColor?: string;
+    mode?: SceneMode;
+    verticalExaggeration?: number;
+    verticalExaggerationRelativeHeight?: number;
+    vr?: boolean;
+    light?: LightProperty;
+    shadow?: ShadowProperty;
+    imageBasedLighting?: ImageBasedLighting;
+  };
+} & Omit<CoreViewerProperty, "scene">;
