+++ web/src/beta/features/PluginPlayground/Viewer/hooks.ts
@@ -1,5 +1,6 @@
+import { ViewerProperty } from "@reearth/beta/features/Editor/Visualizer/type";
 import { Camera } from "@reearth/beta/utils/value";
-import { MapRef, ViewerProperty } from "@reearth/core";
+import { MapRef } from "@reearth/core";
 import { config } from "@reearth/services/config";
 import {
   MutableRefObject,
