+++ web/src/beta/features/Visualizer/hooks.ts
@@ -1,5 +1,6 @@
+import { ViewerProperty } from "@reearth/beta/features/Editor/Visualizer/type";
 import { Camera } from "@reearth/beta/utils/value";
-import { ViewerProperty, ComputedFeature, ComputedLayer } from "@reearth/core";
+import { ComputedFeature, ComputedLayer } from "@reearth/core";
 import { useCallback, useLayoutEffect, useMemo, useRef, useState } from "react";
 
 import { useVisualizerCamera } from "./atoms";
